Fix a couple of portability issues.

+++ src/OptPme.C
@@ -623,7 +623,7 @@ void OptPmeCompute::initializeOptPmeCompute () {
   sp_zstorage = new float [nzlines * zlen];  
 
   printf ("%d: Allocate %d bytes of storage for %d PME Pencils\n", CkMyPe(), 
-	  nzlines * zlen * sizeof(double), nactive);
+	  nzlines * zlen * (int)sizeof(double), nactive);
   
   assert (zline_storage != NULL);
   double * zblock = zline_storage;
